move Command Sender and Receiver in SocketCommunication
