+++ Rover/main.py
@@ -1,46 +1,23 @@
 import socket
 import sys
 import os
-from dis import Instruction  # Assurez-vous d'avoir correctement importé la classe Instruction
-
+from dis import Instruction
 current_dir = os.path.dirname(os.path.abspath(__file__))
 root_dir = os.path.abspath(os.path.join(current_dir, '..'))
 sys.path.append(root_dir)
 
 from Domain.MissionRover.rover import Rover
 from Domain.MissionRover.instruction import Instruction
-from Domain.Communication.CommunicationAbstraction import CommandSender, CommandReceiver
-from Communication.ProtocolCommunication import MyCommunicationProtocol
+from SocketCommunication.ProtocolCommunication import MyCommunicationProtocol
+# from Domain.Communication.CommunicationAbstraction import CommandReceiver, CommandSender
+from SocketCommunication.CommandReceiverMissionControl import CommandReceiverMissionControl
+from SocketCommunication.CommandSenderMissionControl import CommandSenderMissionControl
 from Missions.marsMission import *
-
-class MyCommandSender(CommandSender):
-    def send_command(self, protocol, rover, coords):
-        data_to_send = protocol.encode(rover, *coords)
-        protocol.client_socket.sendall(data_to_send)
-
-class MyCommandReceiver(CommandReceiver):
-    def receive_command(self, protocol):
-        # Réception de la longueur des données en tant que préfixe
-        data_length_bytes = protocol.client_socket.recv(4)
-        data_length = int.from_bytes(data_length_bytes, byteorder='big')
-
-        # Réception des données
-        data = b""
-        while len(data) < data_length:
-            chunk = protocol.client_socket.recv(data_length - len(data))
-            if not chunk:
-                raise Exception("Connexion interrompue avant la fin de la réception des données.")
-            data += chunk
-
-        # Décodage des données
-        instructions_received = data.decode('utf-8')
-        instructions = Instruction(instructions_received, True)
-        return instructions
 
 def main():
     # Initialisation des objets de communication
-    sender = MyCommandSender()
-    receiver = MyCommandReceiver()
+    sender = CommandSenderMissionControl()
+    receiver = CommandReceiverMissionControl()
     protocol = MyCommunicationProtocol(sender, receiver)
 
     # Définition de l'adresse IP et du port du serveur auquel se connecter
